\name{EnergyOnlineCPM-package}
\alias{EnergyOnlineCPM-package}
\alias{EnergyOnlineCPM}
\docType{package}
\title{
EnergyOnlineCPM Package
}
\description{
EnergyOnlineCPM provides users a new function for nonparametric Phase II MSPC.
}
\details{
This package provides users a new function for nonparametric Phase II multiple multivariate change points detection.
}
\author{
Yafei Xu <yafei.xu@hotmail.de>
}
\references{
[1] Szekely, G. J. & Rizzo, M. L. (2004). Testing for equal distributions in high dimension, InterStat.

[2] Hawkins, D. M., Qiu, P. & Kang, C. W. (2003). The changepoint model for statistical process control, Journal of Quality Technology.

[3] Xu, Y. F. (2017). Reference manual: An R package "EnergyOnlineCPM".
 
see URL: https://sites.google.com/site/EnergyOnlineCPM/
}
\keyword{ Phase II Statistical Process Control}
\keyword{ Change Point Model }
\keyword{ Energy Statistic }
\keyword{ High Dimensional Time Series Monitoring }


