\name{EnergyOnlineCPM-package}
\alias{EnergyOnlineCPM-package}
\alias{EnergyOnlineCPM}
\docType{package}
\title{
An R Package for Phase II multiple change points detection for high dimensional time series.
}
\description{
This package is used for Phase II multiple change points detection for high diemensional time series.
}
\details{
\tabular{ll}{
Package: \tab EnergyOnlineCPM\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2017-02-14\cr
License: \tab GPL (>= 2) \cr
}
}
\author{
Yafei Xu <yafei.xu@hu-berlin.de>
}
\keyword{ Phase II Statistical Process Control}
\keyword{ Change Point Model }
\keyword{ Energy Statistic }
\keyword{ High Dimensional Time Series Monitoring }